pyproject.toml
[project]
name = "jigsaw-puzzle"
version = "0.1.0"
description = "Jigsaw Puzzle Helper System"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "boto3>=1.28.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.1.1",
    "mangum>=0.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "moto>=5.1.15",
    "mypy>=1.8.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = false  # DynamoDB等の戻り値がAnyなのは許容
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false  # 外部ライブラリの影響を受けやすいため無効化
warn_no_return = true
check_untyped_defs = true
strict_equality = true

# 外部ライブラリとテストコードの型チェックを緩和
[[tool.mypy.overrides]]
module = [
    "boto3.*",
    "botocore.*",
    "moto.*",
    "PIL.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.services.image_processor"
# 将来実装のため、型チェックをスキップ
ignore_errors = true

[[tool.mypy.overrides]]
module = "tests.*"
# テストコードは型チェックを緩和
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "app.api.*"
# FastAPIルート関数はresponse_modelで型指定されているため緩和
disallow_untyped_defs = false
